requirements.txt
typer==0.15.2
pathlib == 1.0.1
google-auth == 2.38.0
google-oauth2-tool == 0.0.3
google-auth-oauthlib == 1.2.1
google-api-python-client == 2.166.0